Rename infernece to inspect
