+++ app/ui/src/features/inspect/inference.module.scss
