+++ app/ui/src/features/inspect/stream/stream-container.tsx
