+++ app/ui/src/features/inspect/stream/stream.tsx
