+++ app/ui/src/features/inspect/toolbar.tsx
